- Fixes MSPDTI-99 and MSPDTI-100. - Abstracted federation loop detection to handle duplicate request IDs in general. - Refactored and split federation loop SoapUI test case into two: a test case for testing duplicate request IDs from multiple, sequential requests with the same request ID and a test case for testing actual federation loops. - Implemented perioidic request cache clean up using Spring task scheduling (replaced request removal upon transaction completion). - Renamed federationLoop HpdPlusError type to duplicateRequestId.

+++ pdti-server/src/main/java/gov/hhs/onc/pdti/ws/handler/DirectoryRequestCacheDescriptor.java
@@ -3,9 +3,9 @@ package gov.hhs.onc.pdti.ws.handler;
 public interface DirectoryRequestCacheDescriptor {
     public String toCacheBuilderSpecString();
 
-    public int getCleanUpInterval();
+    public long getCleanUpInterval();
 
-    public void setCleanUpInterval(int cleanUpInterval);
+    public void setCleanUpInterval(long cleanUpInterval);
 
     public int getConcurrencyLevel();
